requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
tensorflow
joblib
matplotlib
seaborn
tqdm
streamlit
limes
nltk
pickle